refactor: split up ai.py file (#263)
* split up ai module

* decouple AI template from builder class and rename

* combine prepare function and add k8s kwargs

* add missing parameters

* add test for s2i builder

* disable capsys to see s2i log during run

* mock out s2i system call

* mock out s2i and docker calls

* add test parameters and mock out aws calls

* integrate deployment properties into image_builder

* fix kwargs for local deploy

* raise error for unknown orchestrator when using str

* import AITemplate in `meta_ai` module
remove broad dockerizer import

* fix logging for system command and add test

* transform dict to list
