+++ superai/meta_ai/__init__.py
@@ -1,3 +1,3 @@
 from .ai import AI
+from .ai_template import AITemplate
 from .base import BaseModel
-from .dockerizer import *
