+++ superai/utils/system_command.py
@@ -14,8 +14,16 @@ def system(command: str) -> int:
         Return code of the command
     """
     logger.info(f"Running '{command}'")
-    process = subprocess.Popen(shlex.split(command), stdout=subprocess.PIPE)
-    (out, _) = process.communicate()
+    process = subprocess.Popen(shlex.split(command), stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    with process.stdout:
+        try:
+            for line in iter(process.stdout.readline, b""):
+                logger.info(line.decode("utf-8").strip())
+        except subprocess.CalledProcessError as e:
+            logger.error(f"{str(e)}")
+            raise
+    process.wait()
     if process.returncode != 0:
-        raise subprocess.CalledProcessError(process.returncode, command, output=out)
+        logger.error(f"Command '{command}' failed with return code {process.returncode}")
+        raise subprocess.CalledProcessError(process.returncode, command, output=process.stdout, stderr=process.stderr)
     return process.returncode
